backend/requirements.txt
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
langdetect==1.0.9
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
psycopg2-binary==2.9.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.2
wsproto==1.2.0
